+++ AdventOfCode2019/Day7.cs
@@ -12,12 +12,12 @@ namespace AdventOfCode2019
         public (string, string?) GetAns(string[] input)
         {
             var program = Intcode.ParseProgram(input[0]);
-            var settings = new List<int>
+            var settings = new List<long>
             {
                 0, 1, 2, 3, 4
             };
 
-            var feedbackSettings = new List<int>
+            var feedbackSettings = new List<long>
             {
                 5, 6, 7, 8, 9
             };
@@ -25,18 +25,18 @@ namespace AdventOfCode2019
             return (GetMaxPower(program, settings).ToString(), GetMaxPower(program, feedbackSettings).ToString());
         }
 
-        private static IEnumerable<int> FeedBackHelper(int[] value)
+        private static IEnumerable<long> FeedBackHelper(long[] value)
         {
             while (true) yield return value[0];
             // ReSharper disable once IteratorNeverReturns
         }
 
-        private static int GetMaxPower(IList<int> program, List<int> settings)
+        private static long GetMaxPower(IList<long> program, List<long> settings)
         {
-            var maxPower = int.MinValue;
+            var maxPower = long.MinValue;
             do
             {
-                var currentFeedBack = new[] {0};
+                var currentFeedBack = new[] {0L};
                 var lastAmplifier =
                     Intcode.RunProgram(program, new[] {settings[0]}.Concat(FeedBackHelper(currentFeedBack)));
                 for (var i = 1; i < 5; i++)
